requirements.txt
requests
beautifulsoup4
pandas
seaborn
matplotlib
scikit-learn
umap-learn
wikipedia-api
plotly